Rename ec2instance to ec2vm
